Fix: determine changes correctly for (new) PRs
For new PRs prevRevision is set to null, which causes the plugin to create a diff of files changes based on all commits ever done to the repo.
Fix this for GitHub PullRequests by using the commit of the target branch as the revision to compare against. This way new PRs will no longer incorrectly trigger for not included regions and existing PRs will always trigger for every commit done to them if any commit in the PR/branch touches any files in the included regions